Created initial css layout for landing page.
Added css styles. There are still things to do for the landing page. Media queries are needed to make the
landing page mobile and tablet friendly. Future refactoring of code needs to occur to work on writing clean
code. the new branch is ready to merge and will be merged. Any future additions will be added on future branches.